requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.0.6
django-cors-headers==4.7.0
djangorestframework==3.15.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
